split pool code into seperate namespace
