+++ QSB/ItemSync/ItemManager.cs
@@ -19,18 +19,6 @@ namespace QSB.ItemSync
 			QSBWorldSync.Init<QSBWarpCoreItem, WarpCoreItem>();
 			QSBWorldSync.Init<QSBNomaiConversationStoneSocket, NomaiConversationStoneSocket>();
 			QSBWorldSync.Init<QSBNomaiConversationStone, NomaiConversationStone>();
-			foreach (var streaming in Resources.FindObjectsOfTypeAll<NomaiRemoteCameraStreaming>())
-			{
-				streaming.gameObject.AddComponent<CustomNomaiRemoteCameraStreaming>();
-			}
-			foreach (var camera in Resources.FindObjectsOfTypeAll<NomaiRemoteCamera>())
-			{
-				camera.gameObject.AddComponent<CustomNomaiRemoteCamera>();
-			}
-			foreach (var platform in Resources.FindObjectsOfTypeAll<NomaiRemoteCameraPlatform>())
-			{
-				platform.gameObject.AddComponent<CustomNomaiRemoteCameraPlatform>();
-			}
 		}
 
 		public static IQSBOWItem GetObject(OWItem unityObject)
