+++ QSB/PoolSync/CustomNomaiRemoteCameraPlatform.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-namespace QSB.ItemSync
+namespace QSB.PoolSync
 {
 	internal class CustomNomaiRemoteCameraPlatform : NomaiShared
 	{
