+++ QSB/QSB.csproj
@@ -134,8 +134,8 @@
     <Compile Include="Instruments\QSBCamera\CameraManager.cs" />
     <Compile Include="Instruments\QSBCamera\CameraMode.cs" />
     <Compile Include="Instruments\InstrumentsManager.cs" />
-    <Compile Include="ItemSync\CustomNomaiRemoteCameraPlatform.cs" />
-    <Compile Include="ItemSync\CustomNomaiRemoteCameraStreaming.cs" />
+    <Compile Include="PoolSync\CustomNomaiRemoteCameraPlatform.cs" />
+    <Compile Include="PoolSync\CustomNomaiRemoteCameraStreaming.cs" />
     <Compile Include="ItemSync\Events\DropItemEvent.cs" />
     <Compile Include="ItemSync\Events\DropItemMessage.cs" />
     <Compile Include="ItemSync\Events\MoveToCarryEvent.cs" />
@@ -144,7 +144,7 @@
     <Compile Include="ItemSync\ItemManager.cs" />
     <Compile Include="ItemSync\Patches\ItemPatches.cs" />
     <Compile Include="ItemSync\SocketEventType.cs" />
-    <Compile Include="ItemSync\CustomNomaiRemoteCamera.cs" />
+    <Compile Include="PoolSync\CustomNomaiRemoteCamera.cs" />
     <Compile Include="ItemSync\WorldObjects\IQSBOWItem.cs" />
     <Compile Include="ItemSync\WorldObjects\IQSBOWItemSocket.cs" />
     <Compile Include="ItemSync\WorldObjects\QSBNomaiConversationStone.cs" />
@@ -177,6 +177,8 @@
     <Compile Include="Player\PlayerEntanglementWatcher.cs" />
     <Compile Include="Player\PlayerMapMarker.cs" />
     <Compile Include="Player\PlayerSyncObject.cs" />
+    <Compile Include="PoolSync\Patches\PoolPatches.cs" />
+    <Compile Include="PoolSync\PoolManager.cs" />
     <Compile Include="QSBInputManager.cs" />
     <Compile Include="QSBNetworkLobby.cs" />
     <Compile Include="Patches\QSBPatch.cs" />
